src/server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=src/server/requirements.txt src/server/requirements/requirements.in
#
amqp==5.1.1
    # via kombu
arrow==1.2.3
    # via -r src/server/requirements/requirements.in
asgiref==3.7.2
    # via
    #   -r src/server/requirements/requirements.in
    #   django
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via jsonschema
billiard==4.1.0
    # via celery
bleach[css]==6.0.0
    # via django-post-office
boto3==1.26.159
    # via
    #   -r src/server/requirements/requirements.in
    #   django-ses
boto3-stubs==1.26.159
    # via -r src/server/requirements/requirements.in
botocore==1.29.159
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.29.159
    # via boto3-stubs
celery==5.3.1
    # via
    #   -r src/server/requirements/requirements.in
    #   celery-redbeat
    #   flower
celery-redbeat==2.1.1
    # via -r src/server/requirements/requirements.in
certifi==2023.5.7
    # via
    #   -r src/server/requirements/requirements.in
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   -r src/server/requirements/requirements.in
    #   cryptography
chardet==5.1.0
    # via -r src/server/requirements/requirements.in
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==41.0.1
    # via
    #   -r src/server/requirements/requirements.in
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r src/server/requirements/requirements.in
    #   odfpy
    #   python3-openid
    #   social-auth-core
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.0.0
    # via -r src/server/requirements/requirements.in
django==4.1.9
    # via
    #   -r src/server/requirements/requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-braces
    #   django-configurations
    #   django-cors-headers
    #   django-extensions
    #   django-import-export
    #   django-jazzmin
    #   django-post-office
    #   django-redis
    #   django-ses
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   social-auth-app-django
django-allauth==0.54.0
    # via -r src/server/requirements/requirements.in
django-braces==1.15.0
    # via -r src/server/requirements/requirements.in
django-cache-url==3.4.4
    # via -r src/server/requirements/requirements.in
django-configurations==2.4.1
    # via -r src/server/requirements/requirements.in
django-cors-headers==4.1.0
    # via -r src/server/requirements/requirements.in
django-extensions==3.2.3
    # via -r src/server/requirements/requirements.in
django-import-export==3.2.0
    # via -r src/server/requirements/requirements.in
django-jazzmin==2.6.0
    # via -r src/server/requirements/requirements.in
django-post-office==3.7.0
    # via -r src/server/requirements/requirements.in
django-redis==5.3.0
    # via -r src/server/requirements/requirements.in
django-ses==3.5.0
    # via -r src/server/requirements/requirements.in
django-storages==1.13.2
    # via -r src/server/requirements/requirements.in
django-templated-mail==1.1.1
    # via
    #   -r src/server/requirements/requirements.in
    #   djoser
djangoql==0.17.1
    # via -r src/server/requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r src/server/requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.2
    # via
    #   -r src/server/requirements/requirements.in
    #   djoser
djoser==2.2.0
    # via -r src/server/requirements/requirements.in
drf-spectacular==0.26.3
    # via -r src/server/requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
flake8==6.0.0
    # via -r src/server/requirements/requirements.in
flower==2.0.0
    # via -r src/server/requirements/requirements.in
gunicorn==20.1.0
    # via -r src/server/requirements/requirements.in
honeybadger==0.15.2
    # via -r src/server/requirements/requirements.in
humanize==4.6.0
    # via flower
idna==3.4
    # via
    #   -r src/server/requirements/requirements.in
    #   requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.3.1
    # via celery
markuppy==1.14
    # via tablib
mccabe==0.7.0
    # via flake8
oauthlib==3.2.2
    # via
    #   -r src/server/requirements/requirements.in
    #   requests-oauthlib
    #   social-auth-core
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
pillow==9.5.0
    # via -r src/server/requirements/requirements.in
ply==3.11
    # via djangoql
prometheus-client==0.17.0
    # via flower
prompt-toolkit==3.0.38
    # via click-repl
psutil==5.9.5
    # via honeybadger
psycopg2-binary==2.9.6
    # via -r src/server/requirements/requirements.in
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via -r src/server/requirements/requirements.in
pyjwt[crypto]==2.7.0
    # via
    #   -r src/server/requirements/requirements.in
    #   django-allauth
    #   djangorestframework-simplejwt
    #   social-auth-core
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   celery
    #   celery-redbeat
python3-openid==3.2.0
    # via
    #   django-allauth
    #   social-auth-core
pytz==2023.3
    # via
    #   -r src/server/requirements/requirements.in
    #   django-post-office
    #   django-ses
    #   djangorestframework
    #   flower
pyyaml==6.0
    # via
    #   drf-spectacular
    #   tablib
redis==4.5.5
    # via
    #   celery-redbeat
    #   django-redis
requests==2.31.0
    # via
    #   -r src/server/requirements/requirements.in
    #   django-allauth
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   django-allauth
    #   social-auth-core
s3transfer==0.6.1
    # via boto3
sentry-sdk==1.26.0
    # via -r src/server/requirements/requirements.in
six==1.16.0
    # via
    #   -r src/server/requirements/requirements.in
    #   bleach
    #   honeybadger
    #   python-dateutil
social-auth-app-django==5.2.0
    # via
    #   -r src/server/requirements/requirements.in
    #   djoser
social-auth-core==4.4.2
    # via
    #   -r src/server/requirements/requirements.in
    #   social-auth-app-django
sqlparse==0.4.4
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
tenacity==8.2.2
    # via celery-redbeat
tinycss2==1.1.1
    # via bleach
tornado==6.3.2
    # via flower
types-awscrt==0.16.19
    # via
    #   botocore-stubs
    #   types-s3transfer
types-s3transfer==0.6.1
    # via boto3-stubs
typing-extensions==4.6.3
    # via
    #   asgiref
    #   dj-database-url
tzdata==2023.3
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   -r src/server/requirements/requirements.in
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
whitenoise==6.5.0
    # via -r src/server/requirements/requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
